Fixed the "log 0 = infinity" bug
Because numpy thinks log of zero is infinity, we were getting that weird error where it said something like "can't take the range of infinity lmao" so now any values less than or equal to zero in our spectrogram are put to a value slightly above zero.

As a sidenote, the reason that this didn't happen with the short audio samples is basically because we got lucky. Since songs are longer, there are periods of silence at the beginning and end of each that likely corresponded with 0. The sound effect I played was so short that there was no time for silence. This is also why we were unable to replicate the error using only microphone input.